+++ packages/components/src/divider/component.tsx
@@ -7,7 +7,7 @@ import { space } from '../utils/space';
 import type { DividerProps } from './types';
 import styles from './style.module.scss';
 
-function UnconnectedDivider(
+export function UnconnectedDivider(
 	props: WordPressComponentProps< DividerProps, 'hr', false >,
 	forwardedRef: ForwardedRef< any >
 ) {
@@ -46,24 +46,6 @@ function UnconnectedDivider(
 
 /**
  * `Divider` is a layout component that separates groups of related content.
- *
- * ```js
- * import {
- * 		__experimentalDivider as Divider,
- * 		__experimentalText as Text,
- * 		__experimentalVStack as VStack,
- * } from `@wordpress/components`;
- *
- * function Example() {
- * 	return (
- * 		<VStack spacing={4}>
- * 			<Text>Some text here</Text>
- * 			<Divider />
- * 			<Text>Some more text here</Text>
- * 		</VStack>
- * 	);
- * }
- * ```
  */
 export const Divider = contextConnect( UnconnectedDivider, 'Divider' );